refactor: rename, remove app (sorry)
