+++ src/app/entities/geo-location-info/geo-location-info.ts
