+++ src/app/entities/summary/summary.interface.ts
@@ -1,6 +1,6 @@
 import {Weather} from '../weather/app.weather';
 import {SerializableClassInterface} from '../serializable/serializable.class.interface';
-import {GeoLocationInfo} from '../geo-location-info/app.geo-location-info';
+import {GeoLocationInfo} from '../geo-location-info/geo-location-info';
 import {ZonesInterface} from '../intensity-zones/intensity-zone.interface';
 
 export interface SummaryInterface extends SerializableClassInterface {
